feat: abstract out useTokensSortedByValue for TokenPicker
