+++ src/lib/web3/hooks/useTokens.ts
@@ -1,5 +1,7 @@
 import BigNumber from 'bignumber.js';
 import { useCallback, useMemo, useState } from 'react';
+import { Coin } from '@cosmjs/proto-signing';
+
 import {
   Token,
   getTokenId,
@@ -15,6 +17,7 @@ import {
   useToken,
   useTokenByDenom,
 } from './useDenomClients';
+import { useUserBankValues } from './useUserBankValues';
 
 const { REACT_APP__CHAIN_ID = '' } = import.meta.env;
 
@@ -187,4 +190,70 @@ export function useTokenValueTotal(
   else {
     return null;
   }
+}
+
+type TokenCoin = Coin & {
+  token: Token;
+  value: BigNumber | undefined;
+};
+
+export function useTokensSortedByValue(tokenList: Token[]) {
+  const allUserBankAssets = useUserBankValues();
+  const allUserBankAssetsByTokenId = useMemo(() => {
+    return allUserBankAssets.reduce<{ [symbol: string]: TokenCoin }>(
+      (acc, asset) => {
+        const symbol = getTokenId(asset.token);
+        if (symbol) {
+          acc[symbol] = asset;
+        }
+        return acc;
+      },
+      {}
+    );
+  }, [allUserBankAssets]);
+
+  const { data: nativeChain } = useNativeChain();
+
+  // define sorting rows by token value
+  const sortByValue = useCallback(
+    (tokenA: Token, tokenB: Token) => {
+      const a = getTokenId(tokenA) || '';
+      const b = getTokenId(tokenB) || '';
+      // sort first by value
+      return (
+        getTokenValue(b).minus(getTokenValue(a)).toNumber() ||
+        // if value is equal, sort by amount
+        getTokenAmount(b).minus(getTokenAmount(a)).toNumber() ||
+        // if amount is equal, sort by local chain
+        getTokenChain(tokenB) - getTokenChain(tokenA) ||
+        // lastly sort by symbol
+        tokenA.symbol.localeCompare(tokenB.symbol)
+      );
+      function getTokenValue(id: string) {
+        const foundUserAsset = allUserBankAssetsByTokenId[id];
+        return foundUserAsset?.value || new BigNumber(0);
+      }
+      function getTokenAmount(id: string) {
+        const foundUserAsset = allUserBankAssetsByTokenId[id];
+        return new BigNumber(foundUserAsset?.amount || 0);
+      }
+      function getTokenChain(token: Token) {
+        if (nativeChain && token.chain.chain_id === nativeChain.chain_id) {
+          return 2;
+        }
+        if (token.ibc) {
+          return 1;
+        }
+        return 0;
+      }
+    },
+    [allUserBankAssetsByTokenId, nativeChain]
+  );
+
+  // sort tokens
+  return useMemo(() => {
+    // sort by USD value
+    // create new array to ensure re-rendering with new reference
+    return [...tokenList].sort(sortByValue);
+  }, [tokenList, sortByValue]);
 }
